--- src/main/java/User.java
@@ -1,26 +1,0 @@
-public class User {
-    private int id;
-    private String name;
-    private String email;
-    private String type;
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getType() {
-        return type;
-    }
-}
