+++ src/main/java/grp4/iotbay/User.java
@@ -1,0 +1,54 @@
+package grp4.iotbay;
+
+public class User {
+    private int id;
+    private String name;
+    private String password;
+    private String email;
+    private String type;
+
+    public User() {
+
+    }
+
+    public User(String name, String email,  String password, String type) {
+        this.name = name;
+        this.email = email;
+        this.type = type;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+    public String getEmail() {
+        return email;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getType() {
+        return type;
+    }
+}
