+++ src/main/webapp/staff-home.jsp
@@ -13,7 +13,10 @@
 <body>
 <h1>Welcome to IoT Stock Management staff member: <%=Name%> </h1>
 <h2><a href="add-product.jsp">Add Product</a></h2>
+<h2><a href="update-product.jsp">Update Product Stock/Description</a></h2>
 <h2><a href="remove-product.jsp">Remove Product</a></h2>
+<h2><a href="LogoutServletMod">Logout</a></h2>
+
 
 </body>
 </html>
